when logging to a file, create it with read permissions
Prior to this patch, the file was *supposed* to be opened with
permissions 664, but go was treating that as a decimal number,
causing the file to be write-only. Prepending a 0 causes the
compiler to interpret it as an octal number and therefore the file
to be created with porper permissions.