avoid potential race condition in ID assignment
Were two notifications to be recieved simultaneously, the ID is supposed
to be incremented atomically such that unique ID's are recieved.
However, in the code prior to this patch, we immediately access notif.Id
even though it may have been mutated *again* since our own atomic
mutation, thus duplicating a notification ID and violating the spec, in
addition to potentially causing some problems down the line if we ever
need to replace that notification.

Becaue atomic.AddUint32 returns the old value, we can re-calculate the
next ID by adding to that and bypassing the above issue.