[tentative] move entry point to cmd/painted/main.go
This also renames the main package to painted and reserves main for the
actual program. Aside from making this more consistent with the rest of
the ecosystem, this should hopefully clean up some of the unclear nature
of visibility.

However: this may not work. Go being go, I'm only referring to the
painted module via its github URL, which does not contain this change.