allow command matching by prefix
This allows, a user to enter, for instance, `ex`, as a shorthand for the
full command `exit`.

For this commit, we introduce a ternary search trie in pkg/trie which
stores keys lexographically ordered by runes. It contains two public
methods, Insert and Search, which do exactly what they sound like.
Search, it should be noted, attempts to match keys *after* the given
prefix, and will return the empty set if a full key with no following
potentials are given. Thus, there's some special logic in CmdLoop to
account for that and attempt to match the given command if it doesn't
match as a prefix.