log actions in extended display
The default format for this is `%b | %a` (or body text followed by a
pipe, followed by a list of actions). The %a sequence is more than a bit
inflexible, and should see some changes before making it into a release.
Other than that, this is well on its way to fully supporting actions!