Revert "[copilot] replace blockUntilModify inotify(2) with poll(2)"
This reverts commit 1983a496b32abb27b8d38e834f46da5526e1af19. Poll was
not actually blocking, causing high CPU usage (100% of a thread at all
times) when reading from a file.

Reverting this commit restores original behavior with inotify instead of
poll, which is a Linux-only syscall as far as I'm aware. Restoring POSIX
compatibility will is a requiremnet before the next release.