use flags package to parse arguments
This replaces the custom parse logic with something that is (hopefully)
less brittle. There are a number of removals here (the obvious one being
the parsing logic), but also the help message and logic to display it,
as flags takes care of all of it.

This also removes support for short options (-i, -o, and -c), and adds
support for non-GNU style `-long options`.