handle unknown escape characters
If the user supplies a format string that contains a sequence beginning
with a % and followed by a character not in { a, b, o, s, i }, it's not
an actual escape sequence. Rather than just ignoring it, painted should
log that it was unknown and insert the literal sequence.

I.e., for the given format string below...

	[%a] %s %j

...painted should render, for example, the following notification:

	[notify-send] lorem ipsum %j

In terms of implementation, this is implemneted as a simulated backup in
the DFA. Perhaps a 'more correct' implementation would involve writing
out the full NFA, but the reconstruction/backup implemented is perfectly
acceptable and much easier to understand given the relative lack of
complexity.

This is part of fixing #10.