fix buffer overread when running actions on an empty queue
On an empty buffer, commands like next, prev, expand, etc, use the queue
interface to select a notification and then to invoke a callback. If
there were no notifications, however, the callback would try to access
the 0'th element of an empty deque. This patch just verifies that the
queue is non-empty before invoking a callback, making calls to
notifications on an empty queue effectively a nop.