model: emit dbus dismissal on removal
When a notification is cleared or removed, the dbus spec requires that
notif daemons send org.freedesktop.Notifications.NotificationClosed.
This was done on clear, but must also be done on removal (as implemented
in this commit).

This extracts the function notification.Dismiss, which makes use of the
global state provided in cf336adc.